Added Bean Validation

+++ src/main/java/com/gustavoalsantos/java_crud_client/services/ClientService.java
@@ -10,6 +10,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.NoSuchElementException;
 
@@ -62,6 +63,10 @@ public class ClientService {
 
     @Transactional
     public void delete(Long id) {
+        // Verificar se o ID existe
+        if (repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         repository.deleteById(id);
     }
